next version

+++ XRDpy/incident/main.py
@@ -1,35 +1,37 @@
 import argparse
 import shutil
 import yaml
-import pickle
 import matplotlib.pylab as plt
 from pathlib import Path
 from datetime import datetime
 import XRDpy.package_params as package
-import XRDpy.incident.macro as imacro
-import XRDpy.incident.plot as iplot
+import XRDpy.incident.macro as macro
+import XRDpy.incident.specular as specular
 
 
 def plot():
     parser = argparse.ArgumentParser(
-        prog="XRD-iplot",
+        prog="plot",
         description="Plot om scan as an image with horizontal integration over pixels, rows: y-pixels, columns: om, color: intensity\nThis will look for data in the latest directory of the user",
         epilog="author: Teddy Tortorici <edward.tortorici@colorado.edu>"
     )
     parser.add_argument("user", help="This will be the name of the directory the files will be looked for")
-    parser.add_argument("-A", "--animate", action="store_true")
+    parser.add_argument("-A", "--animate", help="animate: specify frame-rate in FPS")
     parser.add_argument("-T", "--title")
-    parser.add_argument("-D", "--date", help="Override current date with format YYYYMMDD")
-    parser.add_argument("-N", "--number", help="Override latest directory of the given (or current) date")
-    parser.add_argument("-CW", "--crop_width", help="Override number of pixels in horizontal crop (default=13)")
-    parser.add_argument("-CA", "--crop_above", help="Override number of pixels above the direct beam to keep in crop (default=100).")
-    parser.add_argument("-CB", "--crop_below", help="Override number of pixels below the direct beam to keep in crop (default=20).")
-    parser.add_argument("-CO", "--crop_offset", help="Move the center of the crop to the right (in pixels)")
-    parser.add_argument("-C", "--color", help="Integer value corresponding to different color schemes")
-    parser.add_argument("--dir", help="Specify a specific directory, or CWD for current working directory")
+    parser.add_argument("-M", "--mod", help="modify fit or plot (z0 adjust for om or change std for z)")
+    parser.add_argument("-C", "--crit", help="add critical angles (comma separated)")
+    parser.add_argument("-Z", "--z", action="store_true", help="change to a z-specular scan")
+    parser.add_argument("-D", "--dir", help="Specify a specific directory, or CWD for current working directory")
+    parser.add_argument("-R", "--range", help="set angular range in degrees")
+    parser.add_argument("-B", "--beamwidth", help="set beam width in mm")
+    parser.add_argument("-S", "--save", help="save the plot at a certain DPI")
     args = parser.parse_args()
-    
-    usr_dir = package.directory / args.user / "om_scans"
+
+    if args.z:
+        spec_type = "z"
+    else:
+        spec_type = "om"
+    usr_dir = package.directory / args.user / "{}_scans".format(spec_type)
 
     if args.date is None:
         date = datetime.now()
@@ -57,59 +59,38 @@ def plot():
             directory = Path.cwd()
         else:
             directory = Path(args.dir)
+    
+    if args.range is None:
+        angular_range = 1.5
+    else:
+        angular_range = args.range
+    if args.beamwidth is None:
+        beamwidth = 1.
+    else:
+        beamwidth = args.beamwidth
 
-    try:
-        angles, intensity_data, direct_beam = iplot.load_tiff_data(directory)
-    except FileNotFoundError:
-        print(f"Could not find data from today: {date_name}.")
-        print(f"Looking for data in the latest directory for user: {args.user}...")
-        latest_dir = max(usr_dir.glob(f"*"), key=lambda dir: int(dir.name.split("-")[0]), default=None)
-        if latest_dir is not None:
-            directory = latest_dir
+    if args.z:
+        if args.mod is None:
+            std = float(args.mod)
         else:
-            raise FileNotFoundError(f"Could not find any data for user: {args.user}!")
-        dir_num = max(usr_dir.glob(f"{date_name}-*"), key=lambda d: int(d.name.split("-")[-1]), default=1)
+            std = 4.
+        spec = specular.SpecularZ(directory, angular_range=angular_range, beam_width=beamwidth, standard_deviations=std)
+    else:
+        spec = specular.SpecularOmega(directory, anglular_range=angular_range, beam_width=beamwidth)
+        if args.crit is None:
+            crit = None
+        else:
+            crit = [float(c) for c in args.mod.split(",")]
+        if args.mod is not None:
+            spec.fit(z0=float(args.mod))
+        spec.plot(critical_angle=crit)
 
-        directory = usr_dir / date_name.rstrip("-1")
-        angles, intensity_data, direct_beam = iplot.load_tiff_data(directory)
-
-    print(f"Loaded {len(angles) + 1} files from {directory.as_posix()}")
-    print(f"Found {len(angles)} angles in the scan")
-    print(f"Data shape is: {intensity_data.shape}")
-    print(f"Direct beam data shape is: {direct_beam.shape}")
-
-    x_pixel_width = 13
-    pixels_above = 100
-    pixels_below = 20
-    x_offset = 0
-    if args.crop_width is not None:
-        x_pixel_width = int(args.crop_width)
-    if args.crop_above is not None:
-        pixels_above = int(args.crop_above)
-    if args.crop_below is not None:
-        pixels_below = int(args.crop_below)
-    if args.crop_offset is not None:
-        x_offset = int(args.crop_offset)
+    if args.animate is not None:
+        fps = args.animate
+        fig, ani = spec.animate_tiffs(fps)
     
-    id_c, db_c = iplot.crop_data(intensity_data, direct_beam, x_pixel_width, pixels_above, pixels_below, x_offset)
-
-    color_code = 0
-    if args.color is not None:
-        color_code = int(args.color)
-    
-    if args.animate:
-        fps = 48
-        fig, ani = iplot.animate_tiffs(id_c, fps, log=True, color_scheme=color_code)
-    
-    iplot.plot_tuning(angles, id_c, pixel_size=0.75, log=True, color_scheme=color_code)
-    if args.title is not None:
-        plt.title(args.title)
-
-    data_to_save = {"angles": angles,
-                    "data": id_c}
-    with open(directory / "data.pkl", "wb") as f:
-        pickle.dump(data_to_save, f)
-    print("Saved data in: " + (directory / "data.pkl").as_posix())
+    spec.save(directory)
+    print("Saved data in: " + directory.as_posix())
     
     plt.show()
     
@@ -117,17 +98,22 @@ def plot():
 def move():
     date = datetime.now()
     parser = argparse.ArgumentParser(
-        prog="XRD-imove",
+        prog="move",
         description="Move files from Data to a user folder.",
         epilog="author: Teddy Tortorici <edward.tortorici@colorado.edu>"
     )
+    parser.add_argument("type", help="Either om or z")
     parser.add_argument("user", help="This will be the name of the directory the files will be moved to")
     parser.add_argument("-A", "--append", action="store_true", help="Add this tag to append the latest set of data")
     args = parser.parse_args()
+
+    if args.type.lower() not in ["om", "z", "omr", "zr"]:
+        raise ValueError("Type must be 'om', 'z', 'omr', or 'zr'.")
+    
     with open(package.directory / "config.yaml") as f:
         data_path = Path(yaml.safe_load(f)["data_path"])
-
-    usr_dir = package.directory / args.user / "om_scans"
+    
+    usr_dir = package.directory / args.user / "{}_scans".format(args.type.lower())
     if not usr_dir.exists():
         usr_dir.mkdir(parents=True, exist_ok=True)
     
@@ -155,34 +141,48 @@ def move():
         directory.mkdir(parents=True, exist_ok=True)
         print(f"Making new directory: {directory.as_posix()}")
 
-    scan_glob = data_path.glob("om_scan*.tif")
+    scan_glob = data_path.glob(f"{args.type.lower()}_scan*.tif")
     print(f"Found {len(list(scan_glob))} files to move")
 
-    for tif in data_path.glob("om_scan*.tif"):
+    for tif in data_path.glob(f"{args.type.lower()}_scan*.tif"):
         new_tif = directory / tif.name
         print(new_tif)
         if new_tif.is_file():
             new_tif.unlink()
         shutil.move(tif, new_tif)
 
-def make():
+def make_scan():
     parser = argparse.ArgumentParser(
-        prog="XRD-imacro",
-        description="Create a macro for SAXS to scan through om motor angles to calibrate incident angle to a flat sample.",
+        prog="macro_om",
+        description="Create a macro for SAXS to scan through motor angles to calibrate for GIXS.",
         epilog="author: Teddy Tortorici <edward.tortorici@colorado.edu>"
     )
-    parser.add_argument("start_angle")
-    parser.add_argument("end_angle")
-    parser.add_argument("angle_step")
+    parser.add_argument("type")
+    parser.add_argument("start")
+    parser.add_argument("end")
+    parser.add_argument("step")
+    parser.add_argument("-R", "--relative", action='store_true', help="relative scan from current position")
     # parser.add_argument("-T", "-thickness", help="sample thickness")
     parser.add_argument("-N", "--name", help="add a short tag to the TIFF files generated")
     parser.add_argument("-C", "--clear", action='store_true', help="remove all macro files saved")
     dir = Path.home() / "XRDpy" / "Macros"
     args = parser.parse_args()
+    if args.type.lower not in ["om", "z"]:
+        raise ValueError("Invalid type. Must be 'om' or 'z'.")
     if args.clear:
         for macro in dir.glob("*.txt"):
             macro.unlink()
-    angles = imacro.arange_angles(args.start_angle, args.end_angle, args.angle_step)
     if args.name is None:
         args.name = ""
-    imacro.create_file(dir, angles, args.name)
+    if args.relative:
+        macro.create_rel_file(dir, args.type.lower(), args.start, args.end, args.step, args.name)
+    else:
+        list_to_scan = macro.arange_list(args.start, args.end, args.step)
+        if args.type.lower() == "om":
+            macro.create_om_file(dir, list_to_scan, args.name)
+        elif args.type.lower() == "z":
+            macro.create_z_file(dir, list_to_scan, args.name)
+
+
+
+
